+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # v0.3.0:
 * Added method to estimate optimal delay time `estimate_delay`.
+* Added abstract `AbstractDataset` supertype.
+* Completely changed the way `Reconstruction` is done.
+* Added more methods, interaction and smart indexing related to the (new)
+  `Reconstruction` object. See the new docstring of `reconstruct` for usage.
+* Added a new method that enables the calculation of the maximum Lyapunov
+  exponent from a numeric timeseries (like an experimental recording).
+  This method (called `numericallyapunov`) operates on a reconstruction, **using
+  4 different algorithms**. See the docstring of `numericallyapunov` for
+  more info!
+* Added `double_pendulum` to famous systems.
 
 # v0.2.0: Major Improvements and new features
 + Improved algorithm for deducing best boxsizes (using `min_pairwise_distance`)